ragognize/requirements.txt
torch==2.10.0
mwparserfromhell==0.7.2
pymediawiki==0.7.5
wtpsplit==2.1.7
beautifulsoup4==4.14.3
pandas==2.3.3
langchain==0.3.18
langchain-core==0.3.80
python-dotenv==1.2.1
langchain-google-genai==2.0.11
FlagEmbedding==1.3.5
h5py==3.15.1

# adapters==1.1.0
# transformers @ git+https://github.com/F4biian/transformers-v4.47.1-and-internal-states.git@75df852cadde684443cbf95304256381203491c2
# transformer_heads==0.2.2